scripts/quick_action_requirements.txt
pystray
pillow
keyboard
requests
